Rename to new naming convention
