+++ packages/framework-integration-tests/src/events/product-availability-changed.ts
