+++ packages/framework-integration-tests/src/events/product-created.ts
