+++ packages/framework-integration-tests/src/events/product-deleted.ts
