+++ packages/framework-integration-tests/src/events/product-updated.ts
