+++ packages/framework-integration-tests/src/events/stock-moved.ts
