+++ packages/framework-integration-tests/src/read-models/product-updates-read-model.ts
@@ -1,7 +1,7 @@
 import { ReadModel, Projects } from '@boostercloud/framework-core'
 import { Admin } from '../roles'
 import { ProjectionResult, UUID } from '@boostercloud/framework-types'
-import { Product } from '../entities/Product'
+import { Product } from '../entities/product'
 
 // This is an example read model for a possible admin-exclusive report to show last and previous updates to products
 @ReadModel({
